requirements.txt
asgiref==3.11.0
boolean.py==5.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
coverage==7.13.2
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
Django==6.0
boto3==1.40.74
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
google-auth==2.45.0
idna==3.11
license-expression==30.4.4
Markdown==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
packageurl-python==0.17.6
packaging==25.0
Pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
py-serializable==2.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-decouple==3.8
requests==2.32.5
rich==14.2.0
rsa==4.9.1
sortedcontainers==2.4.0
sqlparse==0.5.4
tomli==2.4.0
tomli_w==1.2.0
tzdata==2025.2
urllib3==2.6.3
whitenoise==6.11.0
